Moved checkToken method within the auth service
